moved stuff over, ported init to renderr
